+++ src/mo_frontend/typing.ml
@@ -240,6 +240,8 @@ let info env at fmt =
 
 let check_deprecation env at desc id depr =
   match depr with
+  | Some ("M0235" as code) ->
+    warn env at code "%s %s is deprecated for caffeine" desc id
   | Some ("M0199" as code) ->
     if !(env.reported_stable_memory) then ()
     else begin
@@ -1348,23 +1350,19 @@ type hole_candidate =
 let suggestion_of_candidate candidate =
   Option.fold ~none:candidate.desc ~some:Suggest.module_name_as_url candidate.module_name_opt
 
-(* All candidates are subtypes of the required type. The "greatest" of
-   these types is the "closest" to the required type. If we can
-   uniquely identify a single candidate that is the supertype of all
-   other candidates we pick it. *)
-let disambiguate_resolutions (candidates : hole_candidate list) =
-  let add_candidate (frontiers : hole_candidate list) (c : hole_candidate) =
-    let rec go (fs : hole_candidate list) = match fs with
+let disambiguate_resolutions (rel : 'candidate -> 'candidate -> bool) (candidates : 'candidate list) =
+  let add_candidate (frontiers : 'candidate list) (c : 'candidate) =
+    let rec go (fs : 'candidate list) = match fs with
       | [] -> [c]
       | f::fs' ->
-         if T.sub c.typ f.typ then
-           if T.eq c.typ f.typ then
+         if rel c f then
+           if rel f c then
              (* c = f, so we keep both *)
              f :: go fs'
            else
              (* c <: f, so f absorbs c *)
              fs
-         else if T.sub f.typ c.typ then
+         else if rel f c then
            (* f <: c, so c absorbs f *)
            go fs'
          else
@@ -1403,7 +1401,7 @@ let resolve_hole env at hole_sort typ =
     | Named lab1 -> lab = lab1
     | Anon _ -> true
   in
-  
+
   let is_matching_typ typ1 = T.sub typ1 typ
   in
   let has_matching_field_typ = function
@@ -1489,11 +1487,14 @@ let resolve_hole env at hole_sort typ =
              candidate.desc mid id mod_desc mid id call_region call_src)
       explicit_terms
   in
+  (* All candidates are subtypes of the required type. The "greatest" of these types is the "closest" to the required type.
+     If we can uniquely identify a single candidate that is the supertype of all other candidates we pick it. *)
+  let disambiguate_holes = disambiguate_resolutions (fun (c1 : hole_candidate) c2 -> T.sub c1.typ c2.typ) in
   match eligible_terms with
   | [term] -> Ok term
   | [] ->
     let (lib_terms, _) = candidates true env.libs is_lib_module in
-    (match if !Flags.implicit_lib_vals then disambiguate_resolutions lib_terms else None with
+    (match if Option.is_some !Flags.implicit_package then disambiguate_holes lib_terms else None with
     | Some term -> Ok term
     | None ->
       Error (List.map suggestion_of_candidate lib_terms,
@@ -1501,7 +1502,7 @@ let resolve_hole env at hole_sort typ =
               renaming_hints)
     )
   | terms -> begin
-     match disambiguate_resolutions terms with
+     match disambiguate_holes terms with
      | Some term -> Ok term
      | None ->
        let terms = List.map (fun term -> term.desc) terms in
@@ -1518,6 +1519,7 @@ let resolve_hole env at hole_sort typ =
 type ctx_dot_candidate =
   { module_name : T.lab;
     path : exp;
+    arg_ty : T.typ;
     func_ty : T.typ;
     inst : T.typ list;
   }
@@ -1528,31 +1530,53 @@ type ctx_dot_candidate =
     matching module could be imported, and reports an ambiguity error
     when finding multiple resolutions.
  *)
+
+(* Does an instantiation for [tbs] exist that makes [t1] <: [t2]? *)
+let permissive_sub t1 (tbs, t2) =
+  try
+    let (inst, c) = Bi_match.bi_match_subs None tbs None [t1, t2] [] in
+    ignore (Bi_match.finalize inst c []);
+    Some inst
+  with _ ->
+    None
+
+let has_matching_self receiver_ty tf = match tf with
+  | T.{ lab = "Self"; typ = T.Typ con; _ } ->
+     (match Cons.kind con with
+      | T.Def(tbs, t') -> permissive_sub receiver_ty (tbs, t') <> None
+      | _ -> false)
+  | _ -> false
+
+let check_can_dot env (exp : Syntax.exp) tys es at =
+  if not env.pre then
+  if Flags.get_warning_level "M0236" <> Flags.Allow then
+  match exp.it, tys, es with
+  | (DotE(obj_exp, id, _), receiver_ty :: tys, e::es) ->
+     if (id.it = "equal" || Lib.String.chop_prefix "compare" id.it <> None) && List.length tys = 1 then ()
+     else
+       let can_dot = match T.normalize obj_exp.note.note_typ with
+         | T.Obj(_, fs) ->
+           List.exists (has_matching_self receiver_ty) fs
+         | _ -> false
+       in
+       if can_dot then warn env at "M0236" "You can use the dot notation `%s.%s(...)` here"
+                         (match Source.read_region e.at with Some s -> s | None -> "<arg0>") id.it
+  | _, _, _ -> ()
+
 let contextual_dot env name receiver_ty =
-  (* Does an instantiation for [tbs] exist that makes [t1] <: [t2]? *)
-  let permissive_sub t1 (tbs, t2) =
-    try
-      let (inst, c) = Bi_match.bi_match_subs None tbs None [t1, t2] [] in
-      ignore (Bi_match.finalize inst c []);
-      Some inst
-    with _ ->
-      None in
-  let has_matching_self tf = match tf with
-    | T.{ lab = "Self"; typ = T.Typ con; _ } ->
-       (match Cons.kind con with
-       | T.Def(tbs, t') -> permissive_sub receiver_ty (tbs, t') <> None
-       | _ -> false)
-    | _ -> false in
+  let has_matching_self tf = has_matching_self receiver_ty tf in
   let has_matching_self_type (_, (_, fs)) = List.exists has_matching_self fs in
-  let is_matching_func = function (* TODO: normalize first *)
-    | T.{ lab; typ = T.Func (_, _, tbs, first_arg::_, _) as typ; _ } when lab = name.it ->
+  let is_matching_func field =
+    if not (String.equal field.T.lab name.it) then None
+    else match T.normalize field.T.typ with
+    | T.Func (_, _, tbs, first_arg::_, _) as typ ->
       (match permissive_sub receiver_ty (tbs, first_arg) with
-       | Some inst -> Some (typ, inst)
-       | _ -> None)
+        | Some inst -> Some (T.open_ inst first_arg, typ, inst)
+        | _ -> None)
     | _ -> None in
   let find_candidate in_libs (module_name, (module_ty, fs)) =
     List.find_map is_matching_func fs |>
-      Option.map (fun (func_ty, inst) ->
+      Option.map (fun (arg_ty, func_ty, inst) ->
         let path = {
           it = DotE({
               it = module_exp in_libs env module_name;
@@ -1562,24 +1586,30 @@ let contextual_dot env name receiver_ty =
           at = name.at;
           note = empty_typ_note }
         in
-        { module_name; path; func_ty; inst }) in
+        { module_name; path; func_ty; arg_ty; inst }) in
   let candidates in_libs xs f =
     T.Env.to_seq xs |>
       Seq.filter_map f |>
       Seq.filter has_matching_self_type |>
       Seq.filter_map (find_candidate in_libs) |>
       List.of_seq in
+  (* All candidate functions accept supertypes of the required type as their first arguments.
+     The "smallest" of these types is the closest to the required type. *)
+  let disambiguate_candidates = disambiguate_resolutions (fun c1 c2 -> T.sub c2.arg_ty c1.arg_ty) in
   match candidates false env.vals is_val_module with
-  | [oc] -> Ok oc
+  | [c] -> Ok c
   | [] ->
     (match candidates true env.libs is_lib_module with
-    | [oc] when !Flags.implicit_lib_vals -> Ok oc
+    | [c] when Option.is_some !Flags.implicit_package -> Ok c
     | lib_candidates ->
-      Error (List.map (fun candidate -> Suggest.module_name_as_url candidate.module_name) lib_candidates)
-    )
-  | ocs ->
-    let candidates = List.map (fun oc -> oc.module_name) ocs in
-    error env name.at "M0224" "overlapping resolution for `%s` in scope from these modules: %s" name.it (String.concat ", " candidates)
+      match if Option.is_some !Flags.implicit_package then disambiguate_candidates lib_candidates else None with
+      | Some c -> Ok c
+      | None ->  Error (List.map (fun candidate -> Suggest.module_name_as_url candidate.module_name) lib_candidates))
+  | cs -> match disambiguate_candidates cs with
+    | Some c -> Ok c
+    | None ->
+      let candidates = List.map (fun c -> c.module_name) cs in
+      error env name.at "M0224" "overlapping resolution for `%s` in scope from these modules: %s" name.it (String.concat ", " candidates)
 
 let rec infer_exp env exp : T.typ =
   infer_exp' T.as_immut env exp
@@ -2581,18 +2611,17 @@ and infer_callee env exp =
   match exp.it with
   | DotE(exp1, id, note) -> begin
     match try_infer_dot_exp env exp.at exp1 id ("a function", is_func_typ) with
-    | Ok t -> infer_exp_wrapper (fun _ _ -> t) T.as_immut env exp, None
+    | Ok t ->
+      infer_exp_wrapper (fun _ _ -> t) T.as_immut env exp, None
     | Error (t1, mk_e) ->
       match contextual_dot env id t1 with
       | Error [] ->
-         let e = mk_e () in
-         let sug = "\nHint: If you're trying to use a contextual call you need to import the corresponding module." in
-         Diag.add_msg env.msgs Diag.{ e with text = e.text ^ sug }; raise Recover
+         Diag.add_msg env.msgs (mk_e ()); raise Recover
       | Error suggestions ->
          let e = mk_e () in
          let sug = Format.sprintf "\nHint: Did you mean to import %s?" (String.concat " or " suggestions) in
          Diag.add_msg env.msgs Diag.{ e with text = e.text ^ sug }; raise Recover
-      | Ok { module_name; path; func_ty; inst; } ->
+      | Ok { module_name; path; func_ty; inst; _ } ->
         note := Some path;
         if not env.pre then
           check_exp env func_ty path;
@@ -2670,7 +2699,9 @@ and infer_call env exp1 inst (parenthesized, ref_exp2) at t_expect_opt =
     | _ -> [exp2]
   in
   let t_args, extra_subtype_problems = match ctx_dot with
-    | None -> t_args, []
+    | None ->
+      check_can_dot env exp1 t_args syntax_args at;
+      t_args, []
     | Some(e, t, _id, _inst) -> begin
       match t_args with
       | t'::ts -> ts, [(t, t')]